requirements.txt
annotated-types==0.7.0
argcomplete==3.6.2
attrs==25.3.0
autocommand==2.2.2
Brlapi==0.8.6
btrfsutil==6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
coverage==7.9.1
cryptography==45.0.4
cycler==0.12.1
daemonize==2.5.0
dbus-python==1.4.0
defusedxml==0.7.1
distro==1.9.0
distro-info==1.11
filelock==3.18.0
fonttools==4.58.4
Glances==4.3.1
idna==3.10
iniconfig==2.1.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
kiwisolver==1.4.6
libfdt==1.7.2
libvirt-python==11.4.0
louis==3.34.0
matplotlib==3.10.3
more-itertools==10.7.0
nftables==0.1
nose==1.3.7
numpy==2.3.1
orjson==3.10.18
packaging==25.0
perf==0.1
pillow==11.2.1
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.6.0
protontricks==1.12.1
psutil==7.0.0
pwquality==1.4.5
pyaml==24.12.0
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyGObject==3.52.3
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5_sip==12.17.0
pytest==8.4.1
python-dateutil==2.9.0
python-linux-procfs==0.7.3
pyudev==0.24.3
PyYAML==6.0.2
Reflector==2023.6.28.0.36.1
requests==2.32.4
setproctitle==1.3.5
setuptools==80.9.0
shtab==1.7.2.dev0+g61ec99d.d20241116
six==1.17.0
smbus==1.1
TBB==0.2
termcolor==2.5.0
tldr==3.4.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.5.0
userpath==1.9.2
validate==5.0.9
vdf==4.0
Wand==0.6.11
wheel==0.45.1
win2xcur==0.1.2
